+++ crypto/pem/pem_lib.cc
@@ -487,7 +487,7 @@ int PEM_write_bio(BIO *bp, const char *name, const char *header,
   int retval = 0;
 
   EVP_EncodeInit(&ctx);
-  nlen = strlen(name);
+  nlen = (int)strlen(name);
 
   if ((BIO_write(bp, "-----BEGIN ", 11) != 11) ||
       (BIO_write(bp, name, nlen) != nlen) ||
@@ -495,7 +495,7 @@ int PEM_write_bio(BIO *bp, const char *name, const char *header,
     goto err;
   }
 
-  i = strlen(header);
+  i = (int)strlen(header);
   if (i > 0) {
     if ((BIO_write(bp, header, i) != i) || (BIO_write(bp, "\n", 1) != 1)) {
       goto err;
@@ -584,7 +584,7 @@ int PEM_read_bio(BIO *bp, char **name, char **header, unsigned char **data,
     buf[++i] = '\0';
 
     if (strncmp(buf, "-----BEGIN ", 11) == 0) {
-      i = strlen(&(buf[11]));
+      i = (int)strlen(&(buf[11]));
 
       if (strncmp(&(buf[11 + i - 6]), "-----\n", 6) != 0) {
         continue;
@@ -684,7 +684,7 @@ int PEM_read_bio(BIO *bp, char **name, char **header, unsigned char **data,
     dataB = tmpB;
     bl = hl;
   }
-  i = strlen(nameB->data);
+  i = (int)strlen(nameB->data);
   if ((strncmp(buf, "-----END ", 9) != 0) ||
       (strncmp(nameB->data, &(buf[9]), i) != 0) ||
       (strncmp(&(buf[9 + i]), "-----\n", 6) != 0)) {
